test: move the file: dependency staleness tests into bun-lock.test.ts
